重构文件结构
