+++ README.md
@@ -1,45 +1,2 @@
-# 太极图形课S1作业
-## 背景简介
-这里可以简要描述作业（项目）的基本背景情况，它源自哪里？能够有效解决哪些问题？可以尽情列举它的亮点哦
-
-## 成功效果展示
-- hw10 加了个洒水器
-
-
-![sph_hw10](./sph_hw10/sph_0.gif)
-![sph_hw10](./sph_hw10/sph_2.gif)
-
-
-- hw4 # reference ==> https://www.shadertoy.com/view/WtScDt#
-
-![procedural_animation](./procedural_animation_hw5/Bandlimited_with_antialiasing.gif)
-
-
-- 一个简单的桌球游戏,代码码累了？来打个桌球吧！
-
-
-![table_tennis](./table_tennis/table_tennis.gif)
-
-
-- Games101作业8
-
-
-
-
-![games101hw8](./games101hw8/spring_mode2.gif)
-
-
-
-
-
-
-
-
-
-## 整体结构（Optional）
-```
--LICENSE
--|data
--README.MD
--xxx.py
-```
+# 计算机图形学大作业
+实现了一个台球游戏
