--- sph_hw10/readme.md
@@ -1,2 +1,0 @@
-- I add a rotating sprinkler which spray liquid
-- you can change the size of liquid
